Add support to install script to install BeEF on OpenBSD.
The tricky part is that on OpenBSD, there might be multiple
Ruby versions available in parallel. They are, and the binaries
using each of thse version, suffixed with the version number.
Therefore add a global variable RUBYSUFFIX default to empty,
and when detecting OpenBSD add this suffix to all Ruby commands
from the script.

Also add a simple install_openbsd function just installing
all the system packages.

run the script with "bash install"
and later on run BeEF with "ruby24 beef"