Remove the dependency on therubyracer
therubyracer has long been unmaintained which is causing both security and build
issues to go unfixed. As beef depends on it only to provide a JS runtime for
execjs, we can drop it and require users to install Node.js. execjs
automatically finds the node binary and uses it as a JS runtime.

This fixes #1478, #1045, #1046, #1249, #1374, #1377, #1395, #1396, #1428, #1429
This also fixes #1433, #1502, #1504, #1540, #1545, #1547